2 create push action (#6)
* Zip生成を実装

* GoogleDriveStrageの仮実装と、パス管理やzip管理のアップデート

* Upload機能の実装

* Push機能のUpdateとLICENSE関係更新

* License関係の自動化

+++ IdentifierArchiveCore/Init.cs
@@ -8,7 +8,7 @@ namespace StudioIdGames.IdentifierArchiveCore
 
         public override ActionInfo? Excute(ReadOnlySpan<string> args)
         {
-            if (args.Length < 1)
+            if (args.Length < 2)
             {
                 return new ActionInfo()
                 {
@@ -17,22 +17,45 @@ namespace StudioIdGames.IdentifierArchiveCore
                 };
             }
 
-            var directoryPath = args[0];
-            var directoryInfo = new DirectoryInfo(directoryPath);
+            var settingsFilePath = args[0];
+            var settingFileInfo = new FileInfo($"{settingsFilePath}/{SettingsFile.FileName}");
+            if (!settingFileInfo.Exists)
+            {
+                return new ActionInfo()
+                {
+                    IsError = true,
+                    Message = $"Settings File が見つかりません。({settingFileInfo.FullName})"
+                };
+            }
+            
+            var setting = SettingsFile.FromBytes(File.ReadAllBytes(settingFileInfo.FullName));
 
-            var identifierFileInfo = new FileInfo($"{directoryInfo.FullName}/{IdentifierFile.ArchiveFileName}");
+            var targetName = args[1];
+            var targetPath = $"{setting.TargetBasePath.TrimEnd('/', '\\')}/{targetName}";
+            var targetDirectoryInfo = new DirectoryInfo(targetPath);
+
+            if (!targetDirectoryInfo.Exists)
+            {
+                return new ActionInfo()
+                {
+                    IsError = true,
+                    Message = $"Target Directory が見つかりません。({targetDirectoryInfo.FullName})"
+                };
+            }
+
+            var identifierFileInfo = new FileInfo($"{targetDirectoryInfo.FullName}/{IdentifierFile.ArchiveFileName}");
             if (!identifierFileInfo.Exists)
             {
                 File.WriteAllText(identifierFileInfo.FullName, IdentifierFile.DefaultValue);
             }
 
-            var currentIdentifierFileInfo = new FileInfo($"{directoryInfo.FullName}/{IdentifierFile.CurrentFileName}");
+            var currentIdentifierFileInfo = new FileInfo($"{targetDirectoryInfo.FullName}/{IdentifierFile.CurrentFileName}");
             if (!currentIdentifierFileInfo.Exists)
             {
                 File.WriteAllText(currentIdentifierFileInfo.FullName, IdentifierFile.DefaultValue);
             }
 
-            var ignoreFileInfo = new FileInfo($"{directoryInfo.FullName}/{GitIgnoreFile.FileName}");
+            var ignoreFileInfo = new FileInfo($"{targetDirectoryInfo.FullName}/{GitIgnoreFile.FileName}");
             if (!ignoreFileInfo.Exists)
             {
                 File.WriteAllText(ignoreFileInfo.FullName, GitIgnoreFile.ValueInTargetDirectory);
